Adding terrain, square, and game state models to the client, as well as a method for generating a fake game state.
Summary:
Adding a basic square and terrain model to the client.

CHECKPOINT: Adding a fake game state generator and a game state class.

Adding a test for the getSquareAt method of the GameState.

Adding a false delay to simulate the downloading of the initial game state.

Fixing a subtle bug when (only sometimes!) the resize event would get fired during loading. We use a bool to track when loading finishes, so we know to not draw until that happens.

Merge branch 'master' into T18-basic-square-terrain-model

Reviewers: swsnider

Reviewed By: swsnider

CC: jwall

Maniphest Tasks: T18

Differential Revision: http://d.sil.as/D51