Making the map directive publish the current square that the user is hovering on on the scope.
Summary: Forcing the map directive to have an API, which lets it modify its parent scope. These published values can be listened to by other parts of the system. In this case, the detail view shows the current values of the 'hoverSquare' output of the map directive (called 'hover' in the parent scope).

Reviewers: jwall

Reviewed By: jwall

CC: swsnider

Maniphest Tasks: T18

Differential Revision: http://d.sil.as/D54