Add etcd-backed session store so that backends and frontends can mutually authenticate users
Summary:
Backends need to be able to confirm that users have previously been authenticated by a frontend. The easiest way to do this is to share the session store. Since this needs to be strongly consistent (it's session storage after all), define a new
session storage type that is backed by an etcd cluster.

Reviewers: jwall

Reviewed By: jwall

Differential Revision: http://d.sil.as/D76