Added order view
Summary:
Made the fake game state generator more deterministic in its randomness. Also, added the currently logged-in player as a param to the fake game state generator.

Updating angular to 1.2.5 for its wonderful bugfixes.

Removing the 'center' API from the <map> directive, as it made no sense to have a different mechanism for changing the map's selection and the map's view, which came up as an issue when the order list was attempting to modify the selection.

Switching around the selection logic to be a little more sane. This should let the order view have a lot of visibility and control over what is selected, hopefully.

Switching around order subclass registration to require that subclasses have an 'id' key that uniquely identifies the constructor. This helps make the logic in the order view nicer, as we can dispatch on the order's type into a bunch of view classes and not need to bleed view details into the model.

Adding some fake order test data.

Adding a basic order view. Yay.

Fix up the tests that I forgot to update earlier.

Adding test for the order view.

Reviewers: jwall, swsnider

Reviewed By: swsnider

Maniphest Tasks: T40

Differential Revision: http://d.sil.as/D102